+++ karma.conf.js
@@ -89,6 +89,8 @@ module.exports = config => {
     // available launchers: https://npmjs.org/browse/keyword/karma-launcher
     browsers: ['ChromeHeadless'],
 
+    browserNoActivityTimeout: 60000,
+
     // Continuous Integration mode
     // if true, Karma captures browsers, runs the tests and exits
     singleRun: true,