+++ CHANGELOG.md
@@ -3,11 +3,17 @@
 
 ## Unreleased
 
-- 🏠 INTERNAL: Remove deprecated dev-dependency `gulp-util`.
-- 🏠 INTERNAL: Upgrade dependencies.
+- 🏠 INTERNAL: Remove deprecated dev-dependency `gulp-util`
+- 🏠 INTERNAL: Upgrade dependencies
 - 🚀 NEW: Extend SassDoc [groups](http://sassdoc.com/configuration/#groups)
   setting to allow nesting groups in named categories --
   [#228](https://github.com/oddbird/sassdoc-theme-herman/issues/228)
+- 🚀 NEW: Add support for CSS `@font-face` `local()` src in font previews --
+  [#173](https://github.com/oddbird/sassdoc-theme-herman/issues/173)
+- 🚀 NEW: Add support for embedded data-URI (`data:...`) font previews --
+  [#197](https://github.com/oddbird/sassdoc-theme-herman/issues/197)
+- 💥 BREAKING: Remove `@font` annotation "formats" option (`{woff, woff2}`) --
+  replaced by `formats` key in font Sass map
 
 
 ## 1.0.1: 2018-01-03